Complete project
This project is a little less complex than some of my other recent
projects. This was because I wanted to make the focus of this project
learning how to make a site using webpack, and to challenge myself
to build a site from nothing but the DOM and one starting div.